Mockex (mockex v0.1.0)

Mockex is a lightweight mocking library for elixir. Create simple mocks for external/third-party calls or mock websocket servers with custom listeners and responses.

Functions

add_listeners(list)
(macro)

Macro to add custom listeners for the mock websocket server incoming messages and define mock responses.

push_to_ws(ws_pid, message)

@spec push_to_ws(pid(), tuple()) :: :noconnect | :nosuspend | :ok

Push a message to the mock websocket server.

setup_mock(mocks, list)
(macro)

Macro to setup several mocks before each test case.

start_mock_ws(listener)

Start a mock websocket server.

use_mock(mock_module, opts \\ [], mocks, list)
(macro)

Macro used to define a mock. Provide the Module, function to be mocked, and the expectation. Then, make function calls as you normally would.

Examples

test "mock the String.at/2 function" do
  use_mock String,
    at: fn _, _ -> "mock response here" end do
    assert String.at("hello", 2) == "mock response here"
  end
end

use_mocks(mocks, list)
(macro)

Macro used to define several mock modules before each test. Provide the Module, function to be mocked, and the expectation. Then, make function calls as you normally would.

Examples

test "mock testing several modules and their functions" do
  use_mocks([
    {Map, [],
    [
      get: fn %{}, "key" -> "value" end
    ]},
    {String, [],
    [
      reverse: fn _ -> :reversed end,
      length: fn _ -> :some_length end
    ]}
  ]) do
    assert Map.get(%{}, "key") == "value"
    assert String.reverse(3) == :reversed
    assert String.length(3) == :some_length
  end
end